Refactor some base pixel stuff out of app.js
