+++ JSVisualizer/src/RobotData.js
@@ -1,0 +1,7 @@
+function Position(x, y, heading) {
+	this.x = x;
+	this.y = y;
+	this.heading = heading;
+}
+
+export {Position};
