+++ run_influx.sh
@@ -1,1 +1,1 @@
-docker run -p 8086:8086 -v influxdb:/var/lib/influxdb influxdb
+docker run --name influxdb -p 8086:8086 -v influxdb:/var/lib/influxdb influxdb
